[flow] fill in gaps in server status
Summary:
Server status information is useful for a couple reasons
* It informs users what workload is keeping the server busy at every point in time.
* It can be queried through our internal logging to determine what process is keeping LSP requests queued.

Before this diff we did not always do a great job updating the status so that it accurately describes what was going on in the server. As a result during initialization the status was often stuck at:
* `Server is initializing (Loading saved state 100/100 (100%))`
* `Server is initializing (Indexing files 100/100 (100%))`

long after saved state was loaded and files were indexed, and during a large recheck (on a file with a large fanout):
* `Server is rechecking (Indexing files 1/1 (100%))`

This diff tries to fill in some of these gaps by adding more server (typechecking) states and new server events to trigger the server into these new states.

Specifically, it adds *states* for
* Restoring heaps. This is the process that happens after we have recovered saved state artifacts.
* Indexing post-process. After indexing each file is done, there is a post processing phase (see Export_search) that takes up a non-negligible time.
* Calculating dependents. This is also known in Types_js as "determine what to recheck". It blocks LSP requests and can be expensive for large fanouts.
* Unaccounted, for work that is mistakenly attributed to the latest state, but which has since ended.

Previously the first three of these states were subsumed in "Loading saved state 100%", "Indexing 100%", and "Indexing 100%", respectively. Especially, in the last case this was quite misleading.

This diff also adds events that trigger the above states:
* `Restoring_heaps_start`
* `Indexing_post_process`, `Indexing_end`
* `Calculating_dependents_start`, `Calculating_dependents_end`

The "stop" equivalent for RestoringHeaps was skipped since the next transition (`Load_libraries_start`) happens immediately after it and so there was no point in transitioning through an Unaccounted phase.

Changelog: [fix] Improved precision of server states, that are visible either in vscode status bar, or through the `flow status` command.

Reviewed By: SamChou19815

Differential Revision:
D51733384

------------------------------------------------------------------------
(from abbb2a06c994005b99c7308b8605fc150083e214)

fbshipit-source-id: f0c783866d3a8e585f2ece6833d023fb2ea0b717