[flow] Use Iterable item type when contextually typing arrays
Summary:
`Array<T>` is a subtype of `Iterable<T>`, however iterables do not allow for numeric access, so previously our contextual typing did not work in this situation:
```
declare const f: (Iterable<string => mixed>) => void;

// Previously: missing-local-annot, Now: OK
 f([x => 1]);
```

In the future, if we want, we can change destructuring to use the same logic, so we can tuple destructure iterables. However, before doing that we need to add an option to disable this as babel loose mode is not compatible with that, similar to our `babel_loose_array_spread` option.

Changelog: [internal]

Some fixes in D51735633

Reviewed By: SamChou19815

Differential Revision: D51725489

fbshipit-source-id: 907d504ee6760e52f6956b0415182405d0192f77