Add start server status to ping result
Summary:
The ping tool helps debug unexpected latency in IDE commands by running a
trivial "ping" command in a loop. Before this diff, we only printed a timestamp
with latency. This diff adds the server status to the output as well.

I've changed the output to JSON-object-per-line, which can be
consumed/transformed easily with jq if needed. This is a bit lazy, honestly, but
I wanted to avoid dealing with escaping for CSV or some other machine-readable
output format.

This might also help with the ongoing efforts to improve server status in the
first place, since we can now observe and record the status using this tool
instead of watching the status bar in VS Code.

Reviewed By: SamChou19815

Differential Revision: D51779094

fbshipit-source-id: 2d69b63626aa88b907c520bdd93408b7abd2c032