[flow][numeric-keys] When doing obj-to-obj subtyping check, treat int-like keys as numbers
Summary:
This sets up the next diff, which adds support for numeric keys in object access and creation. Those numbers are treated as their string equivalents. Here we add support for subtyping of objects with keys which are int-like numbers against `{[number]: X}`.

Our desired behavior, which also best handles existing code, is that `{'1': true}` (in the next diff equivalent to `{1: true}`) is a subtype of both `{[number]: boolean}` and `{[string]: boolean}`. This is pretty odd behavior to model in our type-system, but it is the desired behavior to best handle existing code and patterns.

In this diff, in the context of an object-to-object subtyping check, when checking named properties of one object to the indexer of another, we create those keys as a special `NumericStrKeyT` type. It is a subtype of both numbers and strings. There is no way to annotate such a type. Thankfully, there are also limited ways for a key values to "escape", and in those ways we always just return `string` in order to limit the blast radius of this "odd" behavior. `$Keys` and `Object.values` will always return string, and in an implicit instantiation scenario, e.g. `declare function f<K>({+[K]: mixed}): K; f({1: true})`, we treat the keys as strings as well.

I have attempted to add comments in the code to make this clear.

Changelog: [internal] (next diff is changelog entry)

Reviewed By: SamChou19815

Differential Revision: D51555679

fbshipit-source-id: ca44282575d772de3b9e7afecba09dc085418615