[flow] Never validate libdefs durings init under `experimental.libdef_in_checking=true`
Summary:
This is an obvious thing to do, since that's the point of the flag `experimental.libdef_in_checking=true`. It turns out that this mismatch is also the reason why we appear to have some unsuppressable errors under saved state reinit.

If we are reinit-ing with a saved state with suppressed libdef errors recorded in `local_errors`, but using the saved state under `experimental.libdef_in_checking=true` (what the fixed repro test is doing), we will trigger the issue. Since we have the local_errors, but we never see the suppression, since we never recheck the libdef because libdefs are good by construction if we can reinit from saved state.

This bug is a result of interactions between `experimental.libdef_in_checking=true` and the legacy bad decision to record local errors in saved state. This means that we have to rollout `experimental.libdef_in_checking=true` unconditionally. In the future, when `experimental.libdef_in_checking=true` is unconditionally enabled, we should just kill `local_errors` in saved state.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51832482

fbshipit-source-id: a3480b42c68796d78701ee4c901c396b27c83315