[flow] Fix libdef inclusion logic under `experimental.libdef_in_checking=true` when some libdef files are outside of the root
Summary:
Currently under `experimental.libdef_in_checking=true`, we run the risk of getting into an inconsistent state when some libdef files are outside of the root.

We need to have an invariant that under `experimental.libdef_in_checking=true`, all the libdef files must be included in `Files.mk_next_files`. If the invariant is violated, we might see bugs similar to the ones fixed in D51728352. In this diff, I fixed another one.

I believe we do have correct behavior for path filter with regard to libdef under `experimental.libdef_in_checking=true`. However, we are not setting the entry points of traversal correctly. Currently, the entry points only include root, other includes, and flowlib (builtin libdefs). Libdefs under the root just happen to work, because that are part of the root. Libdefs outside the root won't be traversed, so we have the risk that they are never parsed.

We can easily fix this by including every single libdef files in the starting point list. This would also made it unnecessary to have special logic for flowlib directories.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52000723

fbshipit-source-id: 16daddecf5334223a82c23732d8b0b602c714b19