[flow] Options to batch processing LSP requests
Summary:
One hypothesis I have with slow perf is that we are yielding after every request, even for canceled ones. This is not helpful for when the server is badly queued. This flag is designed to enable the behavior where we batch process requests before yielding until we encounter a request that's not cancelled.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51950664

fbshipit-source-id: f5a2b013b1baed8a414c54bdeb9abfd5f060ca15