[flow] Do not pause for cancelled requests under `experimental.batch_lsp_request_processing=true`
Summary:
In this diff, we will start keep track of more information for each workload. Specifically, we will also record a function to tell us whether a request should be cancelled.

Before we run the workload, we will first run this function under `experimental.batch_lsp_request_processing=true`. If we know this request will be cancelled, then we will not yield after processing this request, which might give the IDE thread more chance to process not-cancelled requests.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51950663

fbshipit-source-id: b5e8dff729a24f1093135a29d1c93c9518d7eaf5