[flow] Allow to skip pausing in LwtLoop
Summary:
This diff adds support for not pausing (yielding) in the `LwtLoop` functor for each iteration of loop. I hypothesize that we might be yielding too often, and Flow server might have better idea to decide whether to yield based on the workload we are handling.

We achieve this through `should_pause` ref. It's default to true, and some implementation can set it to false to avoid `Lwt.pause`

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D51957383

fbshipit-source-id: e42b067146e30571a8904b789a067839fc651b4d