[flow] Update suppress_unused_lazy_2 test for precise dependents calculation
Summary:
This diff prepares for the end of this stack where I make the dependent files
calculation more precise. See that diff for details, but succinctly, the precise
definition of dependents is as follows:

Given:
  A. focused files
  B. dirty dependent files (not relevant here, explained at end of stack)
Compute:
  C. Transitive signature dependents of union(A,B)
  D. Direct implementation dependents of union(A,B,C)
Finally:
  E. All dependent files is union(C,D)

Before this change, cycle_a.js was not a signature dependent of dependency.js.
Thus, when we focus dependency.js, (C) does not contain cycle_a.js, and
therefore (D) does not include cycle_b.js, so we do not find the errors in
cycle_b.js.

This diff ensures that (C) does contain cycle_a.js, which means (D) contains
cycle_b.js, so we do find the expected errors needed for this test.

Reviewed By: SamChou19815

Differential Revision: D52078587

fbshipit-source-id: 4c510281f7bb4bddc74b9756e6b0a3d9f8b53382