[flow] Add configuration for precise dependents
Summary:
This diff adds the flag, but it doesn't do anything yet.

I'm adding a flag here so we can do a rollout, to see the impact of this change
on perf. Assuming nothing goes wrong, I expect to remove this option and the
imprecise calculation quickly.

I've updated tests to always run in precise dependents mode, since the imprecise
method is not worth testing and having tests pass both modes
is also not worth the effort.

Reviewed By: SamChou19815

Differential Revision: D52078592

fbshipit-source-id: c65e3a78e07c72154c5eb9ef4b064cdfbe4578c8