[flow] Update suppress_unused_lazy_3 test for precise dependents calculation
Summary:
This diff prepares for the end of this stack where I make the dependent files
calculation more precise. See that diff for details, but succinctly, the precise
definition of dependents is as follows:

Given:
  A. focused files
  B. dirty dependent files (not relevant here, explained at end of stack)
Compute:
  C. Transitive signature dependents of union(A,B)
  D. Direct implementation dependents of union(A,B,C)
Finally:
  E. All dependent files is union(C,D)

The dependency structure of the files in this test:

```
  A1       B1 B3
   | \     | /
  A2 A3    B2
           |
           B4
```

Before this change, B2 was not a signature dependent of B1. Thus, when we focus
B1, we skip checking B4, because changes to B1 can not possibly affect B4.

This test requires that B4 is checked, so that the suppression in B2 is used.
This diff changes B2 to be a signature dependent of B1, so B4 is checked when B1
changes.

Reviewed By: SamChou19815

Differential Revision: D52078589

fbshipit-source-id: 089341e7a0c6744ebe71a68a55da6b7e4192e9af