[flow] Update suppress_unused_lazy test for precise dependents calculation
Summary:
This diff prepares for the end of this stack where I make the dependent files
calculation more precise. See that diff for details, but succinctly, the precise
definition of dependents is as follows:

Given:
  A. focused files
  B. dirty dependent files
Compute:
  C. Transitive signature dependents of union(A,B)
  D. Direct implementation dependents of union(A,B,C)
Finally:
  E. All dependent files is union(C,D)

Before this change, a.js and b.js were an implementation cycle, but not a cycle
in the sig dependency graph. Because b.js was not a signature dependent of a.js,
(C) above would not contain b.js, and thus D would not include b.js's
dependents.

This diff makes a.js and b.js form a true cycle. Thus, (C) does contains b.js
and (D) contains b.js's dependents, as needed for this test.

Reviewed By: SamChou19815

Differential Revision: D52078590

fbshipit-source-id: d604f054e0ba54c920cacba47d369323db5cabad