[flow][autocomplete][refactor] Move sorting of locals into `append_completion_items_of_autoimports`
Summary:
Continuing the refactor from the previous diff, it's now clear that we always do two potential sorting pass before and after `append_completion_items_of_autoimports`. We can move it into the function, so that the function is guaranteed to produce a sorted result according to some sorting criteria. It will be helpful in the future world, where we fuse together some local results and auto-import results.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52060984

fbshipit-source-id: 0f407fcd2e9261b265577e19ec11b54681aec287