[flow] Demonstrate bug with current normalization for render types
Summary:
Our current method for detecting that we should apply special printing behavior for a render type is incorrect and based on a previous implementation of render types that did not include the React.Element shorthand syntax.

this diff adds a test that exposes the issue with the current approach. The next diff fixes the problem.

Reviewed By: SamChou19815

Differential Revision: D52089014

fbshipit-source-id: 15f3e7531286fee96599442ef1e7c83c3f22aa59