Remove blocking reads/writes from MultiWorkerLwt
Summary:
Before this diff, when sending a job to a worker (or reading the response), we
would first wait for the fd to be writable (or readable, respectively), then
perform blocking operations. Per the removed comments, this was done due to
perceived poor performance at some point in time, although the precise cause of
slowdown is not clear.

This diff removes the waiting and replaces blocking operations with non-blocking
ones. I tried this on a hunch that slow blocking operations were causing low
worker utilization. In theory, if a fd is readable, a subsequent read should not
block. But if it did block, then the server process would be unable to keep
other workers busy.

I also removed the wait operations from Marshal_tools_lwt, as these are
unnecessary. The subsequent read/write operations are already non-blocking and
will wait as needed.

Performance testing suggests that the proposed version performs better.
Furthermore, with this version, we no longer see unpredictable spikes in Parse
and Check time when long-lived workers is enabled.

Reviewed By: SamChou19815

Differential Revision: D52239033

fbshipit-source-id: ebb11eb0ed9624fcac3d8c00ea2917b177371e1e