[flow] Never evaluate ReactDRO in ty_normalizer
Summary:
It should cause any change in printed type, so evaluating it is just waiting time. In addition, evaluating it might result in a type that's not exactly the same as the original type, so it might show some results like `?number = ?number`

It probably won't help perf a lot, but defnitely doesn't hurt.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52277079

fbshipit-source-id: 18e8e3b61b9f1674e6101042efd7d154a1895e4a