[flow] Avoid mapping the entire signature graph
Summary:
I found that we are mapping the entire signature dependency graph just to run a topo-sort on some roots. The nodes we actually use might be much fewer than the entire graph, so it looks very wasteful.

We can avoid this, since our implementation Tarjan's algorithm only requires that we can get a set of members from a map, instead of requiring it to be a map. Therefore, we can make it abstract over a map that only exposes the `find` function, then we can lazily transform the member into a set that we need.

I wasn't able to see a noticeable perf improvement from this though.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52263131

fbshipit-source-id: 2bb109ebf956adf32fc7a2630c081463a9fc123f