[flow][ide] Unbreak find-all-refs and rename on current unsaved file
Summary:
Currently, global find-refs and rename assumes that all changes are saved. If this assumption is not true, then we still return results based on saved content, which will have mismatch with unsaved content in IDE. In this stack, I'm not going to fix everything yet, but just to fix the most common case: user triggers global find-ref/rename from a unsaved file a.js, while all other files don't have unsaved changes.

This diff fixes the broken behavior exposed by the previous diff and re-recorded the test. We will now first compute the local-find-ref results for the current file, and return both the local refs and the definition info to help find refs in other files through recheck. Then we will replace all the recheck-based find-ref results in the current file with the local find ref results, since the local find-ref results are based on potentially unsaved content. For rename, we need to do some additional work to ensure that we are doing AST-diffing on the AST derived from unsaved content as well.

Changelog: [fix][ide] Find-all-references and rename will no longer return stale results on the file that triggers the request. Instead, we will compute the most up-to-date result at the time of the request for the triggering file.

Reviewed By: panagosg7

Differential Revision: D52213460

fbshipit-source-id: 40c121bdd456f82ca3f8ff938714dc61fa6f206a