[flow] set a limit to the size of evaluated hover type
Summary:
After D52283038, we are producing more results as "evaluated" on hover. I noticed that some of the evaluated types were too big. Vscode cannot pretty print these types and renders the entire result as a flat string (no line breaks). To avoid this situation, we set a limit to the size of the evaluated result. This has been set arbitrarily to 100 nodes.

Changelog: [fix] Only print "evaluated" hover type results if the type is of a reasonable size (less than 100 nodes)

Reviewed By: SamChou19815

Differential Revision: D52304128

fbshipit-source-id: 8ed6b8031edd7d26be9acd76238a759ac636b543