[flow][normalizer] do not attempt to evaluate in hover on non-computed input
Summary:
When processing hover type requests, we first compute a normalized version of a type without expanding any EvalT/KeysT, and then, we proceed with re-normalizing the type with the `evaluate_type_destructors` option set to `EvaluateAll`. The second normalization would be performed even when we did not encounter an EvalT/KeysT type in the first time.

This diff changes this so that we only do the second pass if we've found one of these types in the first round.

Except from doing less work during hover type processing, there is one difference in behavior: if in the first round we see a type `{ f: T }` where `T` is an alias for an evaluated type (e.g. a `$ReadOnly`), before, we would evaluate the `$Readonly` and return the result in the second pass. Now we will not run the second pass, because the first one only encountered the alias, not the underlying EvalT.

Implementation is simple: we add a field in the normalizer monad state.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D52453231

fbshipit-source-id: dfcbf308d1886afc200bb9ad80148d1805a78287