[flow] distinguish prop map ids based on origin of types
Summary:
Object types originating from annotations can be generated either during the type-sig phase, typically when the type is imported from a different file (type_sig_merge.ml), or during checking, when the type is defined in the same file (type_annotation.ml).

Object type properties are stored in a map indexed by keys that are either some numerical id, or (for toplevel definitions) the abstract location of the object annotation. The idea behind using the aloc as a key is to enable reuse of the same id for faster obj-to-obj checks regardless of where the type was generated (type-sig or check).

The tests added in the previous diff show a bug caused by allowing both phases to write to the property maps for the same aloc-id. The bug causes a violation of the property that a `FullyResolved` type only contains fully-resolved types deeply in its payload. The problematic situation is possible in the scenario where recursive types A and B that include each other in their bodies are defined in files a.js and b.js. The following sequence will cause a type appearing under a `FullyResolved` constructor to have unresolved parts:
* check phase starts computing type A (which may included unresolved types)
* type A includes imported type B (type-sig), which is lazily computed
* imported type B includes type A' (type-sig), which is the same type as A but imported into file B
* check phase writes unresolved props of A to the prop maps
* imported type B (a fully-resolved type) is eventually forced and computes type A', which reuses the existing entry for its location in prop maps that includes unresolved types.

This diff fixes this issue by making the keys to property maps sensitive to the phase when the object props are generated in the case where an abstract location is used as a key (see changes in source_or_generated_id.ml). This way a key generated during type-sig will always point to property types also generated during type-sig.

For the purposes of equality check of prop-maps, we make `equal_id` only take the location part into account. This way we maintain the performance benefit of using alocs as prop map ids.

Note: A restriction that led to this approach is that we need both type_sig_merge and type_annotation to produce types. The former since that's the only way we can deal with imported types and the latter since we rely on this process for error reporting and typed-AST population. This makes it harder to use one of the two types as the representative for all phases.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision:
D52384387

------------------------------------------------------------------------
(from b8109624803ae4c3fba1d5eeba72ffd93447542e)

fbshipit-source-id: 09008ab6cf372833b687732026dc6b94041e9677