[flow][type-sig] Fix out of order visiting issue of typeof with targs
Summary:
Type sig requires us to visit all the locations in order to make ALoc work.

Currently, we are not doing the right thing with typeof with type arguments, and we will crash with out of order error. Since the feature is unsupported, it's not well tested so the bug was hidden for a while. In the next diff, I will add support for typeof with targs, and the issue is exposed.

Visiting it in order is relatively straightforward. This diff mostly moves stuff around and pass enough information, so that targs can be parsed inside `typeof` after all the nominal name stuff.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D52525224

fbshipit-source-id: 8d1a710ea05b10ed5d7a4944c2938370c3c11831