[flow][website][RFC] Delete support for very old flow.js
Summary:
We only show 80 most recent versions of flow.js in try-flow playground. This means that for a while now, we will not show the option for anything older than 0.126, even though they are published to npm and will exist forever.

We do not commit to support any version other than the current version, so I think we already very generous in showing 80 versions in try-flow. The only use case for supporting old flow.js version is for us to figure out what's the first version a bad behavior is introduced. For this purpose, the answer of "one year ago" vs "four years ago" don't matter that much.

Therefore, let's just kill these legacy code that have not been exercised in production for quite a while.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D52528104

fbshipit-source-id: 5344c14296828bcef85a81f72ec8b2cd2105758b