[flow] Always use precise dependents
Summary:
The associated rollout did not have a noticeable affect on perf as measured
through telemetry, but we know that it can help a lot in exceptional scenarios
and there is no reason to suspect the behavior is incorrect, so I'm going to
remove the option and unconditionally use the new logic.

My theory for why telemetry was unmoved: This change avoids "initial" fanout,
where a new lazy server is rebasing over changes. For these initial rechecks,
the precise deps are an improvement. However, after the initial recheck, the
existing recheck optimizations (sig hash-based skipping) is able to avoid
re-checking the unnecessary dependents.

Reviewed By: SamChou19815

Differential Revision: D52269839

fbshipit-source-id: 4e8987bd006e181322389161604668dff770f9d6