[flow][refactor][EZ] Eliminate `general` type of functions and classes
Summary:
I'm looking into places where we still rely on unresolved tvars and trying to get rid of them, and landed on this case.

The "general" type was used to power generic escape analysis. But generic escape analysis is dead since LTI, so we no longer need this. The only place where we need a different type than the locally inferred function type is for potentially overloaded functions, so I make it optionally and derive it locally most of the time, and renamed the parameter accordingly.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52552904

fbshipit-source-id: 9f1bfb40c7d1ffb69a342b4889389bba41b03a6e