[flow] Improve server-worker communication protocol around canceled requests
Summary:
When we send results to the server from a worker, we send two payloads: the
result data first, then perf measurements.

Before this diff, when a worker sees a cancelation request, it would send
the string `"anything"` followed by the string `"anything"` instead of data and
measurements respectively. This is obviously unsafe, since the memory
representation of the string is incompatible, but we rely on the server never
inspecting the values returned from a canceled worker -- otherwise we would
crash, likely by segfault.

This diff changes the protocol. The first message sent from the worker is either
`None` when canceled or `Some data`. Only in the second case do we send the
second message with measurements.

This change removes the invariant that a canceled job's responses will always be
ignored. Removing this invariant is a good thing in itself, since we no longer
need to know about it.

This change also prepares for a follow-up, where we allow the server to yield
while waiting on a canceled worker's response. Yielding here makes it possible
for the code path expecting a "valid" response to end up receiving the "junk"
response, and this change makes it possible to detect canceled worker responses
from all code paths. (More details for this will be in the follow-up.)

Reviewed By: SamChou19815

Differential Revision: D52553805

fbshipit-source-id: 30728a7c5b7f03cfbeb17e39550c345b22d01d6c