[flow][autocomplete] use EvaluateNone when normalizing results
Summary:
Autocomplete uses normalization to show the types of the various results inline. These types do not need to have all EvalTs evaluated. We can use the same configuration that we use for hover types.

The only place where we'd want types to be evaluated is in `autocomplete_literals` where we specifically call `Ty_normalizer_flow.expand_literal_union`. In that case, however, we don't need to pass in `EvaluateAll`. Evaluation is handled explicitly though the `force_eval` flag. This diff also fixes the case of `Keys` where we were not passing the `force_eval` flag. Effectively this change makes it equivalent to the  EvalT case.

Similar to {D52283038} this is expected to help with perf of autocomplete commands.

Changelog: [internal]

Reviewed By: SamChou19815

Differential Revision: D52677511

fbshipit-source-id: 53c59ff41520a86babbb2e049ecef7019318286f