[flow][ast] Remove kind from declare module AST
Summary:
I plan to reuse most of the declare module infra for namespace support, and I landed upon a TODO to remove some semantic analysis of module kind analysis error out of parsing phase.

This diff first stops storing this field in the AST, as rest of our analysis doesn't look at it at all! The validation logic is still left in place for now, since the same semantic analysis is not implemented in the type checking phase.

Prettier is not using it for pretty-printing:

https://github.com/prettier/prettier/blob/45ad4668ebc133621c7f94e678ce399cab318068/src/language-js/print/flow.js#L73-L74

so I think the change won't break anything important.

Changelog: [parser] DeclareModule AST will no longer have a kind field.

Reviewed By: panagosg7

Differential Revision: D52675384

fbshipit-source-id: 65134a0313a8ea93325d431cb879234ffaaef791