[flow][refactor][EZ] Only expose module_kind from file_sig for testing
Summary:
This diff makes `File_sig.t` opaque, and expose requires and `module_kind` from separate functions. With this, it becomes very clear that `module_kind` is only needed in testing, so I named the function `expose_module_kind_for_testing`, with the hope of killing it completely in the future.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D52713409

fbshipit-source-id: b19eed37a6d9a96c7f4bb951820a02b5d8b6289e