[flow] Split the export validation part of file_sig into a post inference check
Summary:
In the last diff, it's clearly shown that file_sig currently serves two purposes:

1. find requires for dependency tracking
2. compute export validation errors

These two purposes don't have to be done in a single pass. In fact, it's much better to delay 2 to the type checking phase, so that we don't have some errors that don't affect the module types to be raised that early. It also helps readability, so that we know what's going with the visitor.

Changelog: [internal]

Reviewed By: samwgoldman

Differential Revision: D52713406

fbshipit-source-id: 7c5cc8d86790d5eded04529f540aa02aeb5d7d6d