[flow] Hook info in funtype
Summary: This builds a `hook` property into `funtype`s, for use in recording whether a function is a hook or not. This field can be a `HookDecl`, which will correspond to a particular `hook useFoo()` declaration and is unique by an id--in future diffs this will prevent you from casting `useState as typeof useEvilState`, for example. There is also `HookAnnot`, which corresponds to a declared annotation that is a hook (syntax tbd), and `NonHook`, which is what it sounds like. Finally, there is `AnyHook`, which is the equivalent of `any` for places where hook-ness doesn't matter (afaik, this is only for type hints).

Reviewed By: jbrown215

Differential Revision: D51233883

fbshipit-source-id: 1378ed043897e5f3086a4675896f9b7db8af3870