[flow] Config options for compatibility with "hooklike" regular function syntax
Summary:
This diff starts the process of building a "compatibility mode" between existing Flow and a world where hook syntax and component syntax are the only way to express React primitives. This diff only builds a config option for this; later diffs will make functions named like hooks (`useX`) immune to some of the rules of React that we enforce.

This diff also cleans up the component syntax config options. Previously, there were a few different modes for `experimental.component_syntax`, but now we only support "parsing" and "typing", so let's just represent this as a bool. (For this version, we'll still parse config options using the old format, so we don't break on rollout)

Reviewed By: SamChou19815

Differential Revision: D52851864

fbshipit-source-id: efde4f71d88fe2504cdad163472d1a91567d9d45