[flow] Check compatibility between hooks and nonhooks
Summary:
This diff builds the subtyping rules for hook functions.

In D51233883 we introduced the four kinds of hookishness: nonhooks, any, declared hooks, and hook annotations. The rules we enforce here is that any-hook is compatible with any kind of function; nonhooks are only otherwise compatible with nonhooks; declared hooks are can flow into hook annotations but can only be flowed into by hook declarations with the same ids, and hook annotations are compatible with hook annotations but can never flow into hook declarations.

To build good error messages for this, this diff also expands on the "explanation" mechanism of error messages by giving UseOp errors an optional explanation that can be printed as well. We use this to provide a UseOp error for incompatibility between hook functions, and follow it up with a sentence explaining the compatibility of hooks.

Reviewed By: SamChou19815

Differential Revision: D51679609

fbshipit-source-id: b2319e1ae9a589741a85879b4003811586a37cd5