[flow] Make hooklike functions have AnyHook kind
Summary:
This diff puts together the setup from the last few in the stack. Here, name defs that are designated as being hooklike, and function declarations with hooklike names, are made to have their types use the `AnyHook` hook flag in the type system rather than the `NonHook` flag. We accomplish this with a type destructor that ultimately flips the flag in FunTs and admits any other value without changing it. We do the same in the type sig, so that exported hooklike functions are treated as AnyHook as well.

The effect of all this is that a hooklike function is treated as neither a hook nor a regular function by the rules of react checker. You can call a hooklike function in a component with a `useX` name (unlike a regular function) and you can call it from regular JS code as well (unlike a hook). In other words, it behaves like all hooks did before we introduced hook syntax and hook checking.

Reviewed By: SamChou19815

Differential Revision: D52855445

fbshipit-source-id: b95c8843f7aab16b7e6cdef20b4b7bdbf489f2ad