[flow] Name def support for hooklike functions
Summary:
In order to treat values defined with `useX` as having special type system properties (i.e. that they are neither treated like hook syntax hooks nor like regular functions), we need to thread naming information through the type system in a way that we don't at present. The best way to do this is in name_def, since this will let us modify the resulting environment value.

There are a couple cases we care about here. We need `const useX = require('useX')` and `const useY: () => number = ...` to be understood as hooklike functions; in the latter case we can augment the `Annotation` name def with a flag indicating whether it's hooklike; we also do so for function values. In the `require` former case we create a new value-level name_def for `HooklikeValue`s. (We could also just include a flag on values; I preferred to keep the value data structure simple, but * shrug *.)

We don't need to do anything here for function declarations themselves, because all the information necessary to determine whether a function declaration is hooklike is available in the function AST node (its name).

Reviewed By: SamChou19815

Differential Revision: D52855443

fbshipit-source-id: 23e65561bd230ab10435f9b407d865f09d0f7dec