[flow][refactor] Eliminate declare_module_ref
Summary:
Since the module_info of declare module is not read in any meaningful way, we should just kill it rather than writing to something that's never read. This diff renames the module_info getter function to be toplevel_module_info and added validation that it's never called when we are in declare module.

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D52846404

fbshipit-source-id: b1ad4a7bebadb295d743aff4d78005675abdce8d