[flow] Do not create empty tvar for string literal types for autocomplete
Summary:
The purpose of these tvars before LTI is to let them get upper bounds through their interactions with surroundings, so that we can provide type aware autocomplete. After LTI, we now use contextual typing for this purpose. We can't use contextual typing at all to guess the expected type of a string literal annotation, so it doesn't make sense to continue doing that.

(Before LTI, it might be somewhat useful through extreme action-at-distance behavior, like `declare var a: '<autocomplete>'; a as 'foo';`, but we certainly should not try to recreate something like this)

Changelog: [internal]

Reviewed By: gkz

Differential Revision: D52887771

fbshipit-source-id: c21550fc04d3420d17cb34c08c988851e466128b