[flow] Eliminate tvar creation of self type for interfaces
Summary:
Interfaces do not have the concept of this so it doesn't need to go through the complicated code path of tying a "knot" of this type.

This diff simplies it by directly storing the class type in the typed AST, instead of creating a tvar and then unifying it with the class type.

I don't think there is any meaningful difference between storing the t_internal vs t, so I just store the more public t in the typed AST.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52916558

fbshipit-source-id: 457cf8c8640d64cce2571d38100d3dcf50e868d7