[flow] Delay the check that annotated constructor return is void
Summary:
The check doesn't have to be during inference, since it can't change the type. In addition, this helps to ensure that we do not eagarly force lazy OpenT in return_t, which might not be ready at this point when we switch to a fixpoint based knot tying for implicit this parameter in classes.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52970101

fbshipit-source-id: 71b4a2503b4541639ea2a115380a5bb857f323e4