[flow] Move "Exports can only appear at the top level" error detection into module_exports_checker
Summary:
D48284974 moves the check into name_resolver with the hope that we can eliminate these checks in file_sig, but also still do it in a single pass. However, now we have `module_exports_checker` as a full separate pass check, we should just keep these checks in a single place.

My ultimate goal is to:
1. have a single system where we can just linearly scan through a list of statements to calculate the module kind and exports, rather than the status quo of this being implemented in a few other places and entangled with other checks, each with complex visitor overrides
2. have a separate system to flag all invalid CJS exports patterns.

With this diff, the goal 2 is completed.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53018138

fbshipit-source-id: cae468518bcab732f4be38dfa592b481a0903fa8