[flow][tests] Add multiplatform conformance tests involving declare function
Summary:
I want to reuse some of the import_export code to generate a type for module/namespace, but I found that the code is very under-exercised today, since it's only used for the multiplatform feature.

This diff adds more tests involving declare functions, since it's not tested and potentially tricky (type of the declare functions cannot be decided locally due to overloads).

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D53018135

fbshipit-source-id: 48f9669c098b1ebe44f29bb38f182a069aa392fa