[flow][refactor][hard] ThisClassT -> ClassT(ThisInstanceT)
Summary:
`ThisClassT` wraps around an InstanceT, and it is designed to be concretized to ClassT(InstanceT) with the GenericT "this" pinned to the right type. Conceptually, it can be thought as `ClassT(ThisInstanceT)`, where `ThisInstanceT` will fix the wrapped `InstanceT`. This is exactly what this diff tries to do: break it into two types so that `ThisInstanceT` can be used on its own without wrapping a class around it, which will be helpful later in the stack.

In all the places that pattern matches on types, I replaced the old `ThisClassT` cases with `ClassT(ThisInstanceT)`. In places where we concretize `ThisClassT`, I added two cases: concretizing `ClassT(ThisInstanceT)` together and concretizing `ThisInstanceT` alone. Why do we need the former? Because quite a few places don't do concretization on their own, and doesn't expect to see `ClassT` wrapping a not-yet-concretized `ThisInstanceT`.

Changelog: [internal]

Reviewed By: panagosg7

Differential Revision: D52938902

fbshipit-source-id: 42d94927b407f9ceee6d63c6f3b1df64cff24fc1